Split CTrendTests into the two component tests
