+++ include/maths/CTimeSeriesDecompositionDetail.h
@@ -12,11 +12,11 @@
 #include <core/CoreTypes.h>
 
 #include <maths/CCalendarComponent.h>
+#include <maths/CCalendarCyclicTest.h>
 #include <maths/CPeriodicityHypothesisTests.h>
 #include <maths/CSeasonalComponent.h>
 #include <maths/CTimeSeriesDecompositionInterface.h>
 #include <maths/CTrendComponent.h>
-#include <maths/CTrendTests.h>
 #include <maths/ImportExport.h>
 
 #include <boost/ref.hpp>
