+++ lib/maths/unittest/CCalendarCyclicTestTest.h
@@ -4,18 +4,17 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-#ifndef INCLUDED_CTrendTestsTest_h
-#define INCLUDED_CTrendTestsTest_h
+#ifndef INCLUDED_CCalendarCyclicTestTest_h
+#define INCLUDED_CCalendarCyclicTestTest_h
 
 #include <cppunit/extensions/HelperMacros.h>
 
-class CTrendTestsTest : public CppUnit::TestFixture {
+class CCalendarCyclicTestTest : public CppUnit::TestFixture {
 public:
-    void testRandomizedPeriodicity();
-    void testCalendarCyclic();
+    void testAccuracy();
     void testPersist();
 
     static CppUnit::Test* suite();
 };
 
-#endif // INCLUDED_CTrendTestsTest_h
+#endif // INCLUDED_CCalendarCyclicTestTest_h
