+++ lib/maths/unittest/Makefile
@@ -23,8 +23,9 @@ SRCS=\
 	CBjkstUniqueValuesTest.cc \
 	CBootstrapClustererTest.cc \
 	CBoundingBoxTest.cc \
-	CCalendarFeatureTest.cc \
 	CCalendarComponentAdaptiveBucketingTest.cc \
+	CCalendarCyclicTestTest.cc \
+	CCalendarFeatureTest.cc \
 	CCategoricalToolsTest.cc \
 	CChecksumTest.cc \
 	CClustererTest.cc \
@@ -74,6 +75,7 @@ SRCS=\
 	CQuantileSketchTest.cc \
 	CRegressionTest.cc \
 	CRadialBasisFunctionTest.cc \
+	CRandomizedPeriodicityTestTest.cc \
 	CRandomProjectionClustererTest.cc \
 	CSamplingTest.cc \
 	CSeasonalComponentTest.cc \
@@ -88,7 +90,6 @@ SRCS=\
 	CTimeSeriesModelTest.cc \
 	CToolsTest.cc \
 	CTrendComponentTest.cc \
-	CTrendTestsTest.cc \
 	CXMeansOnlineTest.cc \
 	CXMeansOnline1dTest.cc \
 	CXMeansTest.cc \
